handle request failing to parse
When using webread() would sometimes return char instead of a structure. No clear reason, and disappears when reattempted. I therefore added a check that if it does not return a structure then reattempt up to 10 times (it appears to typically just take 1 or 2 reattempts). If it does manage to fail all 10 times (never happened to me yet) then it will display an error message to the screen.